% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VanillaCalendar.R
\name{VanillaCalendar-shiny}
\alias{VanillaCalendar-shiny}
\alias{VanillaCalendarOutput}
\alias{renderVanillaCalendar}
\title{Shiny bindings for VanillaCalendar}
\usage{
VanillaCalendarOutput(outputId, width = "100\%", height = "400px")

renderVanillaCalendar(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended. Use \code{height = 'auto'} with
\code{inputMode = TRUE}.}

\item{expr}{An expression that generates a VanillaCalendar}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\value{
\code{VanillaCalendarOutput()} returns a Shiny output element;
\code{renderVanillaCalendar()} returns a Shiny render function.
}
\description{
Output and render functions for using VanillaCalendar within Shiny
applications and interactive Rmd documents.
}
\examples{
# Only run examples in interactive R sessions
if (interactive()) {
  library(shiny)
  library(VanillaCalendar)

  ui <- fluidPage(
    VanillaCalendarOutput("calendar"),
    verbatimTextOutput("dates")
  )

  server <- function(input, output) {
    output$calendar <- renderVanillaCalendar({
      VanillaCalendar(
        options = list(
          type = "default",
          locale = "en",
          selectionDatesMode = "multiple"
        )
      )
    })
    output$dates <- renderPrint(input$calendar_selected)
  }

  shinyApp(ui, server)
}

}
