% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VanillaCalendar.R
\name{VanillaCalendar}
\alias{VanillaCalendar}
\title{Vanilla Calendar Widget}
\usage{
VanillaCalendar(options, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{options}{A list of options to configure the calendar widget}

\item{width}{The width of the widget. Must be a valid CSS unit (like '100\%', '400px', 'auto') or a number, which will be coerced to a string and have 'px' appended.}

\item{height}{The height of the widget. Must be a valid CSS unit (like '100\%', '400px', 'auto') or a number, which will be coerced to a string and have 'px' appended.}

\item{elementId}{An optional ID for the widget element}
}
\description{
Creates a Vanilla Calendar widget for use in R and Shiny applications.
}
\examples{
# Basic calendar with default settings
if (interactive()) {
  VanillaCalendar(options = list())
}

# Calendar with custom settings
if (interactive()) {
  VanillaCalendar(
    options = list(
      type = "default",
      visibility = "visible",
      settings = list(
        lang = "en",
        selection = list(
          day = "single"
        ),
        visibility = list(
          theme = "light"
        )
      )
    ),
    width = "500px",
    height = "400px"
  )
}

}
